+++ src/main/main.cc
@@ -67,6 +67,12 @@ namespace fs = std::filesystem;
 namespace fs = boost::filesystem;
 namespace errorlib = boost::system;
 #endif
+#if GFLAGS_OLD_NAMESPACE
+namespace gflags
+{
+using namespace google;
+}
+#endif
 
 /*
 * Concurrent queues that communicates the Telemetry Decoder
@@ -86,9 +92,9 @@ int main(int argc, char** argv)
         "See COPYING file to see a copy of the General Public License\n \n");
 
     const std::string gnss_sdr_version(GNSS_SDR_VERSION);
-    google::SetUsageMessage(intro_help);
-    google::SetVersionString(gnss_sdr_version);
-    google::ParseCommandLineFlags(&argc, &argv, true);
+    gflags::SetUsageMessage(intro_help);
+    gflags::SetVersionString(gnss_sdr_version);
+    gflags::ParseCommandLineFlags(&argc, &argv, true);
     std::cout << "Initializing GNSS-SDR v" << gnss_sdr_version << " ... Please wait.\n";
 
 #if CUDA_GPU_ACCEL
@@ -124,7 +130,7 @@ int main(int argc, char** argv)
                             if (!fs::create_directory(p, ec))
                                 {
                                     std::cerr << "Could not create the " << FLAGS_log_dir << " folder. GNSS-SDR program ended.\n";
-                                    google::ShutDownCommandLineFlags();
+                                    gflags::ShutDownCommandLineFlags();
                                     return 1;
                                 }
                         }
@@ -146,7 +152,7 @@ int main(int argc, char** argv)
     catch (const boost::thread_resource_error& e)
         {
             std::cerr << "Failed to create boost thread.\n";
-            google::ShutDownCommandLineFlags();
+            gflags::ShutDownCommandLineFlags();
             std::cout << "GNSS-SDR program ended.\n";
             return 1;
         }
@@ -161,7 +167,7 @@ int main(int argc, char** argv)
                 {
                     std::cerr << "Boost exception: " << boost::diagnostic_information(e) << '\n';
                 }
-            google::ShutDownCommandLineFlags();
+            gflags::ShutDownCommandLineFlags();
             std::cout << "GNSS-SDR program ended.\n";
             return 1;
         }
@@ -176,7 +182,7 @@ int main(int argc, char** argv)
                 {
                     std::cerr << "C++ Standard Library exception: " << ex.what() << '\n';
                 }
-            google::ShutDownCommandLineFlags();
+            gflags::ShutDownCommandLineFlags();
             std::cout << "GNSS-SDR program ended.\n";
             return 1;
         }
@@ -191,7 +197,7 @@ int main(int argc, char** argv)
                 {
                     std::cerr << "Unexpected catch. This should not happen.\n";
                 }
-            google::ShutDownCommandLineFlags();
+            gflags::ShutDownCommandLineFlags();
             std::cout << "GNSS-SDR program ended.\n";
             return 1;
         }
@@ -204,7 +210,7 @@ int main(int argc, char** argv)
               << elapsed_seconds.count()
               << " [seconds]\n";
 
-    google::ShutDownCommandLineFlags();
+    gflags::ShutDownCommandLineFlags();
     std::cout << "GNSS-SDR program ended.\n";
     return return_code;
 }