+++ src/tests/single_test_main.cc
@@ -27,6 +27,12 @@
 #include <gtest/gtest.h>
 #include <iostream>
 
+#if GFLAGS_OLD_NAMESPACE
+namespace gflags
+{
+using namespace google;
+}
+#endif
 
 Concurrent_Queue<Gps_Acq_Assist> global_gps_acq_assist_queue;
 
@@ -36,7 +42,7 @@ DECLARE_string(log_dir);
 
 int main(int argc, char **argv)
 {
-    google::ParseCommandLineFlags(&argc, &argv, true);
+    gflags::ParseCommandLineFlags(&argc, &argv, true);
     try
         {
             testing::InitGoogleTest(&argc, argv);
@@ -54,6 +60,6 @@ int main(int argc, char **argv)
         {
             LOG(WARNING) << "Unexpected catch";
         }
-    google::ShutDownCommandLineFlags();
+    gflags::ShutDownCommandLineFlags();
     return res;
 }